Fixed the determineLegalPawnMove function and solved the issue of it doubling up on the website itself.
Also added an invalidMoveHTML element to show when an invalid move is attempted.

Need to create a simulated game to test various moves so we can see if the determineLegalMove function works further down the line. I found that one turn 4 or so the queen doesnt count legal moves as legal and shows it as invalid.